moves ftp connection into connection package, set passiveMode no longer throws exception
