+++ src/main/java/mercury/server/connections/FtpConnection.java
@@ -1,4 +1,4 @@
-package mercury.server;
+package mercury.server.connections;
 
 import mercury.filesystem.NativeFileSystem;
 import mercury.server.connections.socket.SocketExecutor;
